mychat-st/requirements.txt
gradio>=4.0.0
transformers>=4.30.0
pandas>=2.0.0
pandasai>=2.0.0
langchain-groq>=0.1.0
mysql-connector-python>=8.0.0
torch>=2.0.0
tabulate>=0.9.0
langchain_openai>=0.1.0
python-dotenv>=1.0.0 